refactor: api -> client 폴더 이름 변경
