+++ backend/src/main/java/hyfive/gachita/client/ApiClientConfig.java
@@ -1,4 +1,4 @@
-package hyfive.gachita.api;
+package hyfive.gachita.client;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
