+++ backend/src/main/java/hyfive/gachita/client/geocode/dto/GeoCodeReq.java
@@ -1,4 +1,4 @@
-package hyfive.gachita.api.geocode.dto;
+package hyfive.gachita.client.geocode.dto;
 
 public record GeoCodeReq(
         String request,
