+++ backend/src/main/java/hyfive/gachita/client/geocode/dto/LatLng.java
@@ -1,3 +1,3 @@
-package hyfive.gachita.api.geocode.dto;
+package hyfive.gachita.client.geocode.dto;
 
 public record LatLng(double lat, double lng) {}
