+++ backend/src/main/java/hyfive/gachita/client/kakao/KakaoNaviApiClient.java
@@ -1,8 +1,8 @@
-package hyfive.gachita.api.kakao;
+package hyfive.gachita.client.kakao;
 
-import hyfive.gachita.api.ApiClient;
-import hyfive.gachita.api.kakao.dto.request.DirectionsReq;
-import hyfive.gachita.api.kakao.dto.response.KakaoNaviRes;
+import hyfive.gachita.client.ApiClient;
+import hyfive.gachita.client.kakao.dto.request.DirectionsReq;
+import hyfive.gachita.client.kakao.dto.response.KakaoNaviRes;
 import hyfive.gachita.global.BusinessException;
 import hyfive.gachita.global.ErrorCode;
 import lombok.extern.slf4j.Slf4j;
